[feat/custom-theme-config] allow to merge custom theme to base theme
