+++ packages/utils/src/index.ts
@@ -1,2 +1,3 @@
 export * from './assertion';
 export * from './function';
+export * from './object';
